+++ pkg/services/object/util/key_test.go
@@ -4,14 +4,12 @@ import (
 	"path"
 	"testing"
 
-	"github.com/google/uuid"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
 	"github.com/nspcc-dev/neofs-node/pkg/services/object/util"
 	"github.com/nspcc-dev/neofs-node/pkg/util/state"
 	neofsecdsa "github.com/nspcc-dev/neofs-sdk-go/crypto/ecdsa"
 	neofscryptotest "github.com/nspcc-dev/neofs-sdk-go/crypto/test"
 	"github.com/nspcc-dev/neofs-sdk-go/session"
-	"github.com/nspcc-dev/neofs-sdk-go/user"
 	usertest "github.com/nspcc-dev/neofs-sdk-go/user/test"
 	"github.com/stretchr/testify/require"
 )
@@ -24,8 +22,6 @@ func TestNewKeyStorage(t *testing.T) {
 	require.NoError(t, err)
 	stor := util.NewKeyStorage(&nodeKey.PrivateKey, tokenStor, mockedNetworkState{42})
 
-	owner := usertest.ID()
-
 	t.Run("node key", func(t *testing.T) {
 		key, err := stor.GetKey(nil)
 		require.NoError(t, err)
@@ -33,43 +29,37 @@ func TestNewKeyStorage(t *testing.T) {
 	})
 
 	t.Run("unknown token", func(t *testing.T) {
-		_, err = stor.GetKey(&util.SessionInfo{
-			ID:    uuid.New(),
-			Owner: usertest.ID(),
-		})
+		unknownUser := usertest.ID()
+		_, err = stor.GetKey(&unknownUser)
 		require.Error(t, err)
 	})
 
 	t.Run("known token", func(t *testing.T) {
-		tok := createToken(t, tokenStor, owner, 100)
+		tok := createToken(t, tokenStor, 100)
 
-		key, err := stor.GetKey(&util.SessionInfo{
-			ID:    tok.ID(),
-			Owner: owner,
-		})
+		authUser, err := tok.AuthUser()
+		require.NoError(t, err)
+		key, err := stor.GetKey(&authUser)
 		require.NoError(t, err)
 		require.True(t, tok.AssertAuthKey((*neofsecdsa.PublicKey)(&key.PublicKey)))
 	})
 
 	t.Run("expired token", func(t *testing.T) {
-		tok := createToken(t, tokenStor, owner, 30)
-		_, err := stor.GetKey(&util.SessionInfo{
-			ID:    tok.ID(),
-			Owner: owner,
-		})
+		tok := createToken(t, tokenStor, 30)
+		authUser, err := tok.AuthUser()
+		require.NoError(t, err)
+		_, err = stor.GetKey(&authUser)
 		require.Error(t, err)
 	})
 }
 
-func createToken(t *testing.T, store *state.PersistentStorage, owner user.ID, exp uint64) session.Object {
+func createToken(t *testing.T, store *state.PersistentStorage, exp uint64) session.Object {
 	key := neofscryptotest.ECDSAPrivateKey()
-	id := uuid.New()
-	err := store.Store(key, owner, id[:], exp)
+	err := store.Store(key, exp)
 	require.NoError(t, err)
 
 	var tok session.Object
 	tok.SetAuthKey((*neofsecdsa.PublicKey)(&key.PublicKey))
-	tok.SetID(id)
 
 	return tok
 }